first-agent/pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.7.0",
    "langchain>=1.0.3",
    "langchain-anthropic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-openai>=1.0.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "wikipedia>=1.4.0",
]
